requirements.txt
coverage==6.5.0
dvc==2.28.0
pyyaml==5.3.1
tensorflow==2.10.0